pyproject.toml
[project]
name = "duancuoiki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "insightface>=0.7.3",
    "jinja2>=3.1.6",
    "mediapipe==0.10.14",
    "numpy>=2.4.4",
    "onnxruntime-gpu>=1.24.4",
    "opencv-python>=4.13.0.92",
    "uvicorn>=0.44.0",
]
